[refactor] Move the API module into the utils package
