+++ jira_cli/auth.py
@@ -12,9 +12,9 @@ import requests_oauthlib
 from requests_oauthlib.oauth1_session import TokenRequestDenied
 
 from jira_cli import __title__
-from jira_cli.api import get as api_get
 from jira_cli.exceptions import FailedAuthError, JiraApiError, JiraUnavailable, NoAuthenticationMethod
 from jira_cli.models import ProjectMeta, OAuth
+from jira_cli.utils.api import get as api_get
 
 
 def authenticate(project: ProjectMeta, username: Optional[str]=None, password: Optional[str]=None,
